perf: index rows once instead of rescanning, and clear the remaining lint
Two real rescans on the analysis path, both replaced with a single indexing
pass:
- compare_run_matrix rebuilt each run's raw rows by filtering all ~70k rows of
  the cell once per completed run. Measured at that scale: 33.4ms -> 5.6ms
  (6x), per dataset/variant cell, of which the CLI runs four.
- export_paper_tables sliced det_rows models * (1 + 4 modalities) times; it now
  groups by model and by (model, modality) in one pass.

Also adopts itertools.batched for the manual slice-by-range batching in
completion_job_batches, with strict=False spelled out because a short final
batch is expected. Verified equal to the old slicing for batch sizes 1-7 over
lengths 0-11.

Mechanical cleanups: dict() calls to literals, redundant int() around
math.ceil/round (both already return int), list concatenation to unpacking,
a needless assign-then-return, an unused re-export, os.replace to Path.replace,
nested with statements combined, and next() over single-element slices.

Two exemptions are deliberate and documented in place rather than silently
left failing:
- contextlib.suppress is rejected in run_quality_counters: it measured 5.6x
  slower than the bare handler across the ~70k rows that loop walks.
- PERF401 is off project-wide; it fires on small-N record builders where
  folding a multi-line dict literal into a comprehension reads worse.

ruff check now passes clean across scripts/ and tests/.

+++ scripts/plot_acse_embedding_visualizations.py
@@ -323,7 +323,7 @@ def plot_selected_samples(
     if not selected_ids:
         return
     cols = 2
-    rows = int(math.ceil(len(selected_ids) / cols))
+    rows = math.ceil(len(selected_ids) / cols)
     if components == 3:
         fig = plt.figure(figsize=(12, max(4.2, rows * 4.1)), constrained_layout=True)
         axes_array = np.asarray(
